improve long support in logical functions in imagej.process.function.*
git-svn-id: http://code.imagej.net/svn/imagej/trunk@1624 be23795f-a60e-40d4-ba19-055798872e57

+++ imagej/src/main/java/imagej/process/function/unary/OrUnaryFunction.java
@@ -16,7 +16,7 @@ public class OrUnaryFunction implements UnaryFunction
 	
 	public double compute(double input)
 	{
-		double value = ((int)input) | ((int)constant);
+		double value = ((long)input) | ((long)constant);
 			
 		value = TypeManager.boundValueToType(this.targetType, value);
